增加代线程代理远程读请求, 对art有提升，对alexol没作用，感觉是参数设置以及dram层搜索时间不一样造成的

+++ src/nali_alloc.h
@@ -67,7 +67,7 @@ static inline void init_numa_map() {
     }
 }
 
-static inline int get_numa_id(size_t thread_id) {
+static inline int8_t get_numa_id(size_t thread_id) {
     return numa_map[thread_id];
 }
